// frontend/src/app/api/model/attendance-status.ts
/**
 * Coach Buddy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const AttendanceStatus = {
    Present: 'PRESENT',
    Absent: 'ABSENT',
    Excused: 'EXCUSED'
} as const;
export type AttendanceStatus = typeof AttendanceStatus[keyof typeof AttendanceStatus];